backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
cryptography==41.0.7
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0
celery==5.3.4
redis==5.0.1
hypothesis==6.92.1
supabase==2.3.0
pyjwt==2.8.0
groq==1.0.0
requests==2.32.5
